requirements.txt
numpy
mss
pillow
termcolor
pyserial
requests
pywin32
opencv-python
serial-tool 
pyautogui
keyboard
pyqt5-tools
PyQt5

